requirements.txt
Django==1.10.1
django-cors-headers==2.1.0
djangorestframework==3.6.3
django-push-notifications==1.5.0
apns2==0.2.1
asn1crypto==0.22.0
attrs==16.3.0
cffi==1.10.0
cryptography==1.8.1
django-push-notifications==1.5.0
enum34==1.1.6
gpxpy==1.1.2
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.5
ipaddress==1.0.18
MySQL-python==1.2.5
packaging==16.8
Pillow==3.4.2
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
pyOpenSSL==17.0.0
pyparsing==2.2.0
service-identity==16.0.0
six==1.10.0
yapf